Implement separation of concerns (separate API from App.js)
